refactor: rename bau.ts to executor.ts for clarity
Renamed `src/core/bau.ts` to `src/core/executor.ts` to improve readability and indicate its role in executing commands and extensions. Updated all corresponding imports in `src/webhooks/whatsapp.ts`, `src/webhooks/messenger.ts`, and `src/webhooks/discord.ts`.
