+++ src/core/executor.ts
